CPM-331: split integration test
