+++ tests/back/Pim/Structure/Integration/Query/PublicApi/Family/Sql/AbstractGetRequiredAttributesMasksIntegration.php
@@ -1,0 +1,146 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Akeneo PIM Enterprise Edition.
+ *
+ * (c) 2021 Akeneo SAS (http://www.akeneo.com)
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace AkeneoTest\Pim\Structure\Integration\Query\PublicApi\Family\Sql;
+
+use Akeneo\Pim\Structure\Component\AttributeTypes;
+use Akeneo\Test\Integration\TestCase;
+use Webmozart\Assert\Assert;
+
+abstract class AbstractGetRequiredAttributesMasksIntegration extends TestCase
+{
+    protected function getConfiguration()
+    {
+        return $this->catalog->useMinimalCatalog();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    protected function setUp(): void
+    {
+        parent::setUp();
+        // Note that ecommerce already exists by default
+        $this->givenCurrencyForChannel([['code' => 'ecommerce', 'currencies' => ['USD']]]);
+
+        $this->givenChannels([['code' => 'tablet', 'locales' => ['en_US', 'fr_FR'], 'labels' => ['en_US' => 'tablet', 'fr_FR' => 'Tablette'], 'currencies' => ['USD', 'EUR']]]);
+
+        $this->givenAttributes([
+            ['code' => 'a_non_required_text', 'type' => AttributeTypes::TEXT],
+            // A price because the handling is different than other attribute
+            ['code' => 'a_price', 'type' => AttributeTypes::PRICE_COLLECTION],
+            ['code' => 'a_localizable_non_scopable_price', 'type' => AttributeTypes::PRICE_COLLECTION, 'localizable' => true],
+            // Localizable and Scopable things
+            ['code' => 'a_non_localizable_non_scopable_text', 'type' => AttributeTypes::TEXT],
+            ['code' => 'a_localizable_non_scopable_text', 'type' => AttributeTypes::TEXT, 'localizable' => true],
+            ['code' => 'a_non_localizable_scopable_text', 'type' => AttributeTypes::TEXT, 'scopable' => true],
+            ['code' => 'a_localizable_scopable_text', 'type' => AttributeTypes::TEXT, 'scopable' => true, 'localizable' => true],
+            // Locale specific things
+            ['code' => 'a_non_localizable_non_scopable_locale_specific', 'type' => AttributeTypes::TEXT, 'available_locales' => ['fr_FR']],
+            ['code' => 'a_localizable_non_scopable_locale_specific', 'type' => AttributeTypes::TEXT, 'localizable' => true, 'available_locales' => ['en_US']],
+            ['code' => 'a_non_localizable_scopable_locale_specific', 'type' => AttributeTypes::TEXT, 'scopable' => true, 'available_locales' => ['fr_FR', 'en_US']],
+            ['code' => 'a_localizable_scopable_locale_specific', 'type' => AttributeTypes::TEXT, 'localizable' => true, 'scopable' => true, 'available_locales' => ['fr_FR']],
+        ]);
+    }
+
+    private function givenCurrencyForChannel(array $channelsData): void
+    {
+        $channels = array_map(function ($channelData) {
+            $channel = $this->get('pim_catalog.repository.channel')->findOneBy(['code' => $channelData['code']]);;
+            $this->get('pim_catalog.updater.channel')->update(
+                $channel,
+                [
+                    'currencies' => $channelData['currencies']
+                ]
+            );
+
+            $errors = $this->get('validator')->validate($channel);
+            Assert::count($errors, 0);
+
+            return $channel;
+        }, $channelsData);
+
+        $this->get('pim_catalog.saver.channel')->saveAll($channels);
+    }
+
+    private function givenChannels(array $channelsData): void
+    {
+        $channels = array_map(function ($channelData) {
+            $channel = $this->get('pim_catalog.factory.channel')->create();
+            $this->get('pim_catalog.updater.channel')->update(
+                $channel,
+                [
+                    'code' => $channelData['code'],
+                    'locales' => $channelData['locales'],
+                    'currencies' => $channelData['currencies'],
+                    'category_tree' => 'master'
+                ]
+            );
+
+            $errors = $this->get('validator')->validate($channel);
+            Assert::count($errors, 0);
+
+            return $channel;
+        }, $channelsData);
+
+        $this->get('pim_catalog.saver.channel')->saveAll($channels);
+    }
+
+    protected function givenFamilies(array $familiesData): void
+    {
+        $families = array_map(function ($familyData) {
+            $family = $this->get('pim_catalog.factory.family')->create();
+            $this->get('pim_catalog.updater.family')->update(
+                $family,
+                [
+                    'code' => $familyData['code'],
+                    'attributes'  =>  $familyData['attribute_codes'] ?? [],
+                    'attribute_requirements' => $familyData['attribute_requirements'] ?? [],
+                ]
+            );
+
+            $errors = $this->get('validator')->validate($family);
+            Assert::count($errors, 0);
+
+            return $family;
+        }, $familiesData);
+
+        $this->get('pim_catalog.saver.family')->saveAll($families);
+    }
+
+    private function givenAttributes(array $attributesData): void
+    {
+        $attributes = array_map(function ($attributeData) {
+            $attribute = $this->get('pim_catalog.factory.attribute')->create();
+            $this->get('pim_catalog.updater.attribute')->update(
+                $attribute,
+                [
+                    'code' => $attributeData['code'],
+                    'type' => $attributeData['type'],
+                    'localizable' => $attributeData['localizable'] ?? false,
+                    'scopable' => $attributeData['scopable'] ?? false,
+                    'group' => 'other',
+                    'available_locales' => $attributeData['available_locales'] ?? [],
+                    'decimals_allowed' => $attributeData['type'] === AttributeTypes::PRICE_COLLECTION ? false : null,
+                ]
+            );
+
+            $errors = $this->get('validator')->validate($attribute);
+            Assert::count($errors, 0);
+
+            return $attribute;
+        }, $attributesData);
+
+        $this->get('pim_catalog.saver.attribute')->saveAll($attributes);
+    }
+}
